v0.2 - Fixed output - Added loading_order support

+++ py/globalClasses.py
@@ -55,7 +55,7 @@ class Table:
                         table, variable=lst[i].__getattr__(headerKey), onvalue=1, offvalue=0)
                     self.e.grid(row=i+1, column=j)
                 elif headerKey == 'weight':
-                    lst[i].__setattr__(headerKey, tk.IntVar(value=shipInstanceValue))
+                    lst[i].__setattr__(headerKey, tk.DoubleVar(value=shipInstanceValue))
                     self.e = tk.Entry(
                         table, textvariable=lst[i].__getattr__(headerKey), justify='center')
                     self.e.grid(row=i+1, column=j)
